Create README.md
In this practical, I connected Arduino UNO with the help of a jumper wire with an IR sensor. I connected three pins of jumper wire VCC, GND, and OUT. On the other end of the jumper wire, three pins are connected with Arduino UNO according to their function VCC with 5V, GND with DIGITAL(PWM) GND, and OUT with pin 2  of DIGITAL PWM. Now, as all hardware is connected, I remove my keyboard cable and connect one end of Arduino UNO with the Arduino cable and the other end of the cable with the CPU. Finally, I run the Arduino program.  As the two lights blink the IR sensor is detecting the objects. One light blinking means the IR sensor is not detecting objects. As the program is complete I removed the Arduino cable wire.